refactor(floors): move floor scenes into src/features/floors/
Groups each floor (lobby, platform, architecture, finance, product,
executive) into its own folder under src/features/floors/ alongside
the scene class, info slice, quiz slice, and (for platform) enemy
config. Shared base — LevelScene plus LevelEnemySpawner /
LevelTokenManager / LevelZoneSetup / LevelDialogBindings — now lives
in src/features/floors/_shared/.

Barrels in src/config/quiz and src/config/info still expose the same
public API (QUIZ_DATA, INFO_POINTS, getQuizFor, getInfoPointsFor) but
re-import slices from the new per-floor folders so callers elsewhere
need no changes.

vitest coverage excludes src/features/floors/** and src/systems/sounds/**
to preserve the previous "scenes/sound files skip coverage" behaviour
until Tier C #4 adds the scene coverage thresholds.

Addresses Tier C #3 from docs/architecture-review.md.
