+++ src/features/floors/_shared/LevelScene.ts
@@ -1,15 +1,15 @@
 import * as Phaser from 'phaser';
-import { GAME_WIDTH, GAME_HEIGHT, TILE_SIZE, ELEVATOR_SPEED, FLOORS, FloorId } from '../../config/gameConfig';
-import { LEVEL_DATA, FloorData } from '../../config/levelData';
-import { Player } from '../../entities/Player';
-import { Enemy } from '../../entities/Enemy';
-import { DroppedAU } from '../../entities/DroppedAU';
-import { HUD } from '../../ui/HUD';
-import { ElevatorButtons } from '../../ui/ElevatorButtons';
-import { DialogController } from '../../ui/DialogController';
-import { ProgressionSystem } from '../../systems/ProgressionSystem';
-import { allKeyLabels } from '../../input';
-import type { NavigationContext } from '../NavigationContext';
+import { GAME_WIDTH, GAME_HEIGHT, TILE_SIZE, ELEVATOR_SPEED, FLOORS, FloorId } from '../../../config/gameConfig';
+import { LEVEL_DATA, FloorData } from '../../../config/levelData';
+import { Player } from '../../../entities/Player';
+import { Enemy } from '../../../entities/Enemy';
+import { DroppedAU } from '../../../entities/DroppedAU';
+import { HUD } from '../../../ui/HUD';
+import { ElevatorButtons } from '../../../ui/ElevatorButtons';
+import { DialogController } from '../../../ui/DialogController';
+import { ProgressionSystem } from '../../../systems/ProgressionSystem';
+import { allKeyLabels } from '../../../input';
+import type { NavigationContext } from '../../../scenes/NavigationContext';
 import { LevelEnemySpawner } from './LevelEnemySpawner';
 import { LevelTokenManager } from './LevelTokenManager';
 import { LevelZoneSetup } from './LevelZoneSetup';
