+++ src/features/floors/architecture/ArchitectureTeamScene.ts
@@ -1,6 +1,6 @@
 import * as Phaser from 'phaser';
-import { GAME_HEIGHT, TILE_SIZE, FLOORS } from '../../config/gameConfig';
-import { LevelScene, LevelConfig } from './LevelScene';
+import { GAME_HEIGHT, TILE_SIZE, FLOORS } from '../../../config/gameConfig';
+import { LevelScene, LevelConfig } from '../_shared/LevelScene';
 
 /**
  * Floor 1 — Architecture room (right side of the Platform Team floor).
