+++ src/features/floors/finance/quiz.ts
@@ -1,4 +1,4 @@
-import { QuizDefinition } from './types';
+import { QuizDefinition } from '../../../config/quiz/types';
 
 /**
  * Quizzes for the Finance Team floor (left room of floor 3 / Business).
