+++ src/scenes/products/ProductRoomScene.ts
@@ -1,6 +1,6 @@
 import * as Phaser from 'phaser';
 import { GAME_HEIGHT, GAME_WIDTH, TILE_SIZE, FLOORS } from '../../config/gameConfig';
-import { LevelScene, LevelConfig } from '../floors/LevelScene';
+import { LevelScene, LevelConfig } from '../../features/floors/_shared/LevelScene';
 import { allKeyLabels } from '../../input';
 import type { NavigationContext } from '../NavigationContext';
