+++ vitest.config.ts
@@ -13,10 +13,12 @@ export default defineConfig({
         'src/**/*.test.ts',
         'src/main.ts',
         'src/scenes/**',
+        'src/features/floors/**',
         'src/plugins/**',
         'src/systems/SpriteGenerator.ts',
         'src/systems/sprites/**',
         'src/systems/SoundGenerator.ts',
+        'src/systems/sounds/**',
         'src/systems/MusicGenerator.ts',
         'src/ui/**',
         'src/entities/**',
